Update schema. Refactor listeners.
