+++ src/listeners/index.ts
@@ -1,1 +1,4 @@
 import '~listeners/storage'
+import '~listeners/runtime'
+import '~listeners/tabs'
+import '~listeners/permissions'
