+++ src/listeners/storage/onChanged/index.ts
@@ -1,1 +1,3 @@
 import '~listeners/storage/onChanged/configTextSync'
+import '~listeners/storage/onChanged/matchPatternsSync'
+import '~listeners/storage/onChanged/forwardChanges'
