+++ src/listeners/tabs/index.ts
@@ -1,0 +1,1 @@
+import '~listeners/tabs/onUpdated'
